Problems with drawing circles. Not Fixed

+++ FigureMoving/Classes/Figure.cs
@@ -10,8 +10,8 @@ namespace FigureMoving
 {
     abstract class Figure
     {
-        public int X { get; private set; }
-        public int Y { get; private set; }
+        public float X { get; private set; }
+        public float Y { get; private set; }
 
         public abstract void Draw(Graphics graphics);
